+++ src/main/java/live/smoothing/sensordata/adapter/TopicAdapter.java
@@ -1,6 +1,6 @@
 package live.smoothing.sensordata.adapter;
 
-import live.smoothing.sensordata.dto.SensorTopicResponse;
+import live.smoothing.sensordata.dto.watt.SensorTopicResponse;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
