+++ src/main/java/live/smoothing/sensordata/dto/watt/PowerMetric.java
@@ -1,4 +1,4 @@
-package live.smoothing.sensordata.dto;
+package live.smoothing.sensordata.dto.watt;
 
 import lombok.Builder;
 import lombok.Getter;
