+++ src/main/java/live/smoothing/sensordata/repository/WattRepository.java
@@ -1,6 +1,6 @@
 package live.smoothing.sensordata.repository;
 
-import live.smoothing.sensordata.dto.Watt;
+import live.smoothing.sensordata.dto.watt.Watt;
 
 import java.time.Instant;
 import java.util.List;
